flask/requirements.txt
Flask
mysql-connector-python
apscheduler